Support for heterogeneous nodes (I2C and serial).
Unification of I2C and serial hardware interfaces.
