+++ src/interface/bme280_sensor.py
@@ -25,16 +25,19 @@ class BME280Sensor(I2CSensor):
         return self.data.humidity
 
 
-def discover(*args, **kwargs):
+def discover(config, *args, **kwargs):
+    if 'i2c_helper' not in kwargs:
+        return []
+
     sensors = []
-    config = kwargs['config']
+    i2c_helper = kwargs['i2c_helper']
     supported_bme280_addrs = [0x76, 0x77]
     for addr in supported_bme280_addrs:
         url = I2CSensor.url(addr)
         sensor_config = config.get(url, {})
         name = sensor_config.get('name', url)
         try:
-            sensors.append(BME280Sensor(name, addr, kwargs['i2c_helper']))
+            sensors.append(BME280Sensor(name, addr, i2c_helper))
             print "BME280 found at address {0}".format(addr)
         except IOError:
             print "No BME280 at address {0}".format(addr)
